Error showing , success showing boxes and other mechaninsms, also theme toggle options

+++ BackEnd/Users/views/profile_views.py
@@ -1,8 +1,8 @@
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.permissions import IsAuthenticated
 from rest_framework.parsers import MultiPartParser, FormParser
-from .models import Profile
-from .serializers import UserProfileSerializer
+from Users.models import Profile
+from Users.serializers import UserProfileSerializer
 
 
 class ProfileViewSet(ModelViewSet):
@@ -23,9 +23,11 @@ class ProfileViewSet(ModelViewSet):
         remove_image = request.data.get('remove_image', 'false') == 'true'
 
         if remove_image and profile.profile_image:
-            profile.profile_image.delete()  # Deletes the uploaded image from storage
-            profile.profile_image = profile._meta.get_field(
-                'profile_image').get_default()  # ✅ Set default image
-            profile.save()  # ✅ Save changes to the database
+            default_image = profile._meta.get_field(
+                'profile_image').get_default()
+            if profile.profile_image != default_image:
+                profile.profile_image.delete()  # Deletes the uploaded image from storage
+                profile.profile_image = default_image  # Set default image
+                profile.save()  # Save changes to the database
 
         return super().update(request, *args, **kwargs)
